Move /query folder; update mp req
