+++ desktop/apps/auction/queries/sale.js
@@ -3,7 +3,7 @@ export default function SaleQuery (id) {
     sale(id: "${id}") {
       _id
       associated_sale {
-        ${require('../../../../components/react/auction_block/query.js').default}
+        ${require('../../../components/react/auction_block/query.js').default}
       }
       auction_state
       cover_image {
@@ -22,6 +22,16 @@ export default function SaleQuery (id) {
       is_open
       live_start_at
       name
+      promoted_sale {
+        id
+        name
+        sale_artworks {
+          artwork {
+            is_purchasable
+            is_acquireable
+          }
+        }
+      }
       registration_ends_at
       start_at
       status
