+++ desktop/apps/auction/queries/worksByFollowedArtists.js
