+++ desktop/apps/auction/routes.js
@@ -1,11 +1,11 @@
 import * as actions from 'desktop/apps/auction/actions/artworkBrowser'
 import App from 'desktop/apps/auction/components/App'
 import Articles from 'desktop/collections/articles.coffee'
-import ArticlesQuery from 'desktop/apps/auction/utils/queries/articles'
+import ArticlesQuery from 'desktop/apps/auction/queries/articles'
 import Auction from 'desktop/models/auction.coffee'
-import MeQuery from 'desktop/apps/auction/utils/queries/me'
+import MeQuery from 'desktop/apps/auction/queries/me'
 import React from 'react'
-import SaleQuery from 'desktop/apps/auction/utils/queries/sale'
+import SaleQuery from 'desktop/apps/auction/queries/sale'
 import auctionReducer from 'desktop/apps/auction/reducers'
 import configureStore from 'desktop/components/react/utils/configureStore'
 import footerItems from 'desktop/apps/auction/utils/footerItems'
